Applied alignment changes for pesto and cookies
Updated pesto.html and cookies.html to align with the previous
changes to salad.html and the stylesheet. The italicized descrip sub-
heading was changed to class "subheader". This includes a div class of
descrip, with generic p elements and no <br> betweend description
paragraphs. The notes are placed within <div> elements for organization,
 and a footer style selector is added for index.html.